fix(useMaplibreMap): Make sure only one map instance is created
In react 18, `useEffect` calls are done twice to ensure applications are
reliant to rerenders. This happens only locally, as a prevention
measure. This is why it wasn't broken on deployed version but only
locally. Thx to @dnsos for the tipp. We now make sure to track if a map
was initialized and don't recreate an instance if so. We should make
sure to test against double `useEffect` calls in the future.